Added assertions and docs for pulse waveform

+++ vitis/export/Baremetal/uz/uz_wavegen/uz_wavegen.h
@@ -5,10 +5,9 @@ typedef struct {
 	float amplitude;
 	float frequency_Hz;
 	float DutyCycle;
-
 } wavegen_pulse_settings;
 
-float uz_wavegen_sine(int amplitude, int frequency_Hz);
-float uz_wavegen_sawtooth(int amplitude, int frequency_Hz);
+float uz_wavegen_sine(float amplitude, float frequency_Hz);
+float uz_wavegen_sawtooth(float amplitude, float frequency_Hz);
 float uz_wavegen_pulse(wavegen_pulse_settings pulse);
